fixed sparc atomics; added sparc newlib
added sparc udelay (prototype now in kern/include/timing.h)

+++ kern/arch/sparc/console.c
@@ -10,7 +10,7 @@ cons_init(void)
 void
 cputbuf(const char*COUNT(len) buf, int len)
 {
-	frontend_syscall(RAMP_SYSCALL_write,1,buf,len);
+	frontend_syscall(RAMP_SYSCALL_write,1,(uint32_t)buf,len);
 }
 
 // Low-level console I/O
